Rename placement controller to binding controller
